refactor track list tiles and fix theme issues
